Update imports, migrate OpenIdConnect to its own file
